refactor(sep41): inline test helpers into validator_test.go
